Add ringbuf_buffer_size function to replace RINGBUF_SIZE.
Better to use this function rather than the #define RINGBUF_SIZE, so
that future versions of this data type can support dynamic and/or
different sizes to suit the task at hand.

Performance could be impacted for some of the internal routines that
use the buffer size to compute pointers, but a good compiler should
recognize that ringbuf_buffer_size returns a constant value, anyway.